Refactor N-qubit confusion matrix node and shared readout correction.
Replace confusion_matrix_general with n_qubit_confusion_matrix, rename node
38, consolidate qubit group save logic, and use recover_prepared_probs in
bell state tomography.
